requirements.txt
aiogram==2.25.1
SQLAlchemy==1.4.46
alembic==1.9.4
python-dotenv==0.21.0
requests==2.28.2
Pillow==9.4.0
python-telegram-bot==13.7
gunicorn==20.1.0
secrets==1.0.2
